test-6 added: now using origin
Simultaneous, in-frame movement was causing hitbox and drawing issues,
so now only one input will be read per frame.

Drawing is now in relation to the origin, which will be essential for
rotations to actually work. It also fixed some issues from test-5 in
which boxes were not always staying together (now, every box of the
tetronimo is calculated based off of the origin, so they can not
possibly separate)