test-7 added
test-7 is basically test-6 but with collision detection that works for
all tiles.

Issue right now that needs to be fixed is that when movedown (without
input) happens with movedown from input, it will sometimes ignore
collisions. I'll need to somehow figure out a proper way to only have
one of these happen at a time, or something similar.